# product_manager_api_doc.yml
...
swagger: '2.0'
info:
  version: "1.0.0"
  title: Product Manager - Gestor de Categorias y Productos en una Tienda
  description: |
    </br>
    API para Gestionar Categorias y Productos.
    </br>
    </br>
  contact: 
      name: API Support
      url: https://github.com/jorgeMorfinezM/product_catalog_api/
      email: jorge.morfinez.m@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
host:
  localhost:8080
basePath: "/api/v1.0/manager"
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /product/category:
    post:
      tags:
        - "Manage Categories"
      description:
        Endpoint to manage data for Categories of Products (Create).
      parameters:
        - name: RequestCategory
          in: body
          description: Payload with States data to insert.
          required: true
          schema:
            $ref: '#/definitions/RequestCategory'
      responses:
        200:
          description: Successful response States Register
          schema:
            title: ResponseCategories
            type: array
            items:
              $ref: '#/definitions/ResponseCategories'
        404:
          description: Page Not Found
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request Data Conflict
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
        401:
          description: 401 Unauthorized
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
        500:
          description: Server Error
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
    put:
      tags:
        - "Manage Categories"
      description:
          Endpoint to manage data for Categories of Products (Update).
      parameters:
        - name: RequestCategory
          in: body
          description: Payload with data to updated.
          required: true
          schema:
            $ref: '#/definitions/RequestCategory'
      responses:
        200:
          description: Successful response User updated
          schema:
            title: ResponseCategories
            type: array
            items:
              $ref: '#/definitions/ResponseCategories'
        404:
          description: Page Not Found
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request Data Conflict
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
        401:
          description: 401 Unauthorized
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
        500:
          description: Server Error
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
    delete:
      tags:
        - "Manage Categories"
      description:
          Endpoint to manage data for Categories of Products (Delete).
      parameters:
        - name: RequestCategory
          in: body
          description: Payload with data to inactivate.
          required: true
          schema:
            $ref: '#/definitions/RequestCategory'
      responses:
        200:
          description: Successful response Inactivate User
          schema:
            title: ResponseCategories
            type: array
            items:
              $ref: '#/definitions/ResponseCategories'
        404:
          description: Page Not Found
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request Data Conflict
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
        401:
          description: 401 Unauthorized
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
        500:
          description: Server Error
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
    get:
      tags:
        - "Manage Categories"
      description:
        Get data States.
      parameters:
        - name: Pagination
          in: body
          description: Payload with pagination data to Get.
          required: false
          schema:
            $ref: '#/definitions/Pagination'
      responses:
        200:
          description: Successful response retrieve Categories
          schema:
            title: ResponseCategories
            type: array
            items:
              $ref: '#/definitions/ResponseCategories'
        404:
          description: Page Not Found
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request Data Conflict
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
        401:
          description: 401 Unauthorized
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
        500:
          description: Server Error
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
  /product/category/filter/:
    get:
      tags:
        - "Retrieve Categories by Filters"
      description:
        Get All States registered by filters.
      parameters:
        - name: FilterCategory
          in: body
          description: Payload with filters to find Category.
          required: true
          schema:
            $ref: '#/definitions/FilterCategory'
      responses:
        200:
          description: Successful response
          schema:
            title: ResponseCategories
            type: array
            items:
              $ref: '#/definitions/ResponseCategories'
        404:
          description: Page Not Found
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request Data Conflict
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
        401:
          description: 401 Unauthorized
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
        500:
          description: Server Error
          schema:
            type: array
            items:
              $ref: '#/definitions/Error'
...
definitions:
  Error:
    allOf:
      - $ref: '#/definitions/Error'
      - required:
        - error_code
        - error_message
        type: "object"
        properties:
          error_code:
            type: integer
            format: int32
          error_message:
            type: string
...
  Pagination:
    allOf:
      - $ref: '#/definitions/Pagination'
      - required:
        - offset
        - limit
        type: "object"
        properties:
          offset:
            type: integer
            format: int64
          limit:
            type: integer
            format: int64
  RequestCategory:
    allOf:
      - $ref: '#/definitions/RequestCategory'
      - required:
        - nombre_categoria
        - descrippcion_corta_categoria
        - estatus_categoria
        type: "object"
        properties:
          nombre_categoria:
            type: string
          descrippcion_corta_categoria:
            type: string
          estatus_categoria:
            type: string
  ResponseCategories:
    allOf:
      - $ref: '#/definitions/ResponseCategories'
      - required:
        - id_categoria
        - nombre_categoria
        - descripcion_categoria
        - estatus_categoria
        - creation_date
        - last_update_date
        type: "object"
        properties:
          id_categoria:
            type: integer
            format: int64
          nombre_categoria:
            type: string
          descripcion_categoria:
            type: string
          estatus_categoria:
            type: string
          creation_date:
            type: string
          last_update_date:
            type: string
  FilterCategory:
    allOf:
      - $ref: '#/definitions/FilterCategory'
      - required:
          - nombre_categoria
          - estatus_categoria
        type: "object"
        properties:
          nombre_categoria:
            type: string
          estatus_categoria:
            type: string
...